SSC-0158_CompNuvemArqSer/Code/requirements.txt
Flask==2.0.1
kafka-python==2.0.2
mysql-connector-python==8.0.25
tweepy==3.10.0